feat: UI 컴포넌트 모듈화 완료
- UI 함수들을 컴포넌트로 분리하여 관심사 분리 달성
- 기존 시그니처 유지 원칙으로 안정적인 리팩토링 진행

추가된 컴포넌트:
- Header: 헤더 UI 업데이트
- StockInfo: 재고 정보 UI 업데이트
- OrderSummary: 주문 요약 UI 업데이트
- TotalAndDiscount: 총액 및 할인 정보 UI 업데이트
- CartPrices: 장바구니 가격 UI 업데이트
- ProductSelect: 상품 선택 드롭다운 UI 업데이트

리팩토링 원칙:
- 기존 함수 시그니처 완전 유지
- DOM 직접 접근으로 의존성 최소화
- 단일 책임 원칙(SRP) 적용
- React 마이그레이션 준비
